app/requirements.txt
boto3==1.26.165
psycopg2-binary==2.9.6
pandas==2.0.2